- renamed quota/logging folder
